change int to value

+++ cartesian_vector_param/config.h
@@ -5,4 +5,4 @@
 #endif
 
 // vectors often deal with reals, however ints are better for testing
-using value = int;
+using value = long;
